Move status control to separate component
